requirements.txt
# requirements.txt 
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.1.1
playwright==1.38.0